+++ test/bones_api_test_entities.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/2.3.1
+// BUILDER: reflection_factory/2.3.3
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -20,7 +20,7 @@ typedef __TI<T> = TypeInfo<T>;
 typedef __PR = ParameterReflection;
 
 mixin __ReflectionMixin {
-  static final Version _version = Version.parse('2.3.1');
+  static final Version _version = Version.parse('2.3.3');
 
   Version get reflectionFactoryVersion => _version;
 
@@ -360,12 +360,26 @@ class Account$reflection extends ClassReflection<Account>
   @override
   Map<String, dynamic> getFieldsValues(Account? obj,
       {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'user': obj?.user,
       'userInfo': obj?.userInfo,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(Account? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'user': obj?.user,
+      'userInfo': obj?.userInfo,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -937,6 +951,7 @@ class Address$reflection extends ClassReflection<Address>
   @override
   Map<String, dynamic> getFieldsValues(Address? obj,
       {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'state': obj?.state,
@@ -949,6 +964,25 @@ class Address$reflection extends ClassReflection<Address>
       'closedStores': obj?.closedStores,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(Address? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'state': obj?.state,
+      'city': obj?.city,
+      'street': obj?.street,
+      'number': obj?.number,
+      'latitude': obj?.latitude,
+      'longitude': obj?.longitude,
+      'stores': obj?.stores,
+      'closedStores': obj?.closedStores,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -1422,12 +1456,26 @@ class Photo$reflection extends ClassReflection<Photo> with __ReflectionMixin {
   @override
   Map<String, dynamic> getFieldsValues(Photo? obj,
       {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'data': obj?.data,
       'dataUrl': obj?.dataUrl,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(Photo? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'data': obj?.data,
+      'dataUrl': obj?.dataUrl,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -1966,6 +2014,7 @@ class Role$reflection extends ClassReflection<Role> with __ReflectionMixin {
 
   @override
   Map<String, dynamic> getFieldsValues(Role? obj, {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'type': obj?.type,
@@ -1973,6 +2022,20 @@ class Role$reflection extends ClassReflection<Role> with __ReflectionMixin {
       'value': obj?.value,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(Role? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'type': obj?.type,
+      'enabled': obj?.enabled,
+      'value': obj?.value,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -2529,6 +2592,7 @@ class Store$reflection extends ClassReflection<Store> with __ReflectionMixin {
   @override
   Map<String, dynamic> getFieldsValues(Store? obj,
       {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'name': obj?.name,
@@ -2536,6 +2600,20 @@ class Store$reflection extends ClassReflection<Store> with __ReflectionMixin {
       'owner': obj?.owner,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(Store? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'name': obj?.name,
+      'number': obj?.number,
+      'owner': obj?.owner,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -3137,6 +3215,7 @@ class User$reflection extends ClassReflection<User> with __ReflectionMixin {
 
   @override
   Map<String, dynamic> getFieldsValues(User? obj, {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'email': obj?.email,
@@ -3150,6 +3229,26 @@ class User$reflection extends ClassReflection<User> with __ReflectionMixin {
       'creationTime': obj?.creationTime,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(User? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'email': obj?.email,
+      'password': obj?.password,
+      'address': obj?.address,
+      'roles': obj?.roles,
+      'level': obj?.level,
+      'wakeUpTime': obj?.wakeUpTime,
+      'userInfo': obj?.userInfo,
+      'photo': obj?.photo,
+      'creationTime': obj?.creationTime,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }
@@ -3633,11 +3732,24 @@ class UserInfo$reflection extends ClassReflection<UserInfo>
   @override
   Map<String, dynamic> getFieldsValues(UserInfo? obj,
       {bool withHashCode = false}) {
+    obj ??= object;
     return <String, dynamic>{
       'id': obj?.id,
       'info': obj?.info,
       'idFieldName': obj?.idFieldName,
       'fieldsNames': obj?.fieldsNames,
+      if (withHashCode) 'hashCode': obj?.hashCode,
+    };
+  }
+
+  @override
+  Map<String, dynamic> getJsonFieldsVisibleValues(UserInfo? obj,
+      {bool withHashCode = false}) {
+    obj ??= object;
+    return <String, dynamic>{
+      'id': obj?.id,
+      'info': obj?.info,
+      'idFieldName': obj?.idFieldName,
       if (withHashCode) 'hashCode': obj?.hashCode,
     };
   }