+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.6.9
 
+- `ClassReflectionEntityHandler`:
+  - `getFields`: call `reflection.getJsonFieldsVisibleValues` (avoid issues with `JsonHidden.fields`).
+
 - reflection_factory: ^2.3.2
 
 ## 1.6.8